backend/requirements.txt
fastapi==0.109.0
sqlmodel==0.0.14
uvicorn==0.27.0
python-dotenv==1.0.0
passlib==1.7.4
bcrypt==4.1.2
psycopg[binary]>=3.2.0
